add product detail

+++ src/main/java/com/example/demo/review/dto/ProductReviewInfoResponse.java
@@ -9,5 +9,5 @@ import lombok.Getter;
 @Builder
 public class ProductReviewInfoResponse {
     private Double averageRating;
-    private Long reviewCount;
+    private Integer reviewCount;
 }
